fraud_detection/requirement.txt
django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
